feat(nx-plugin): add force option to update executor and move compare spec
