+++ packages/nx-plugin/src/executors/update-api/schema.d.ts
@@ -1,6 +1,10 @@
 export interface UpdateApiExecutorSchema {
   client: string;
   directory: string;
+  /**
+   * If true, the Client code will be regenerated even if the spec has not changed
+   */
+  force?: boolean;
   name: string;
   plugins: string[];
   scope: string;
